requirements.txt
streamlit==1.28.0
openai==0.28.0
python-dotenv==1.0.0
Pillow>=10.0.0
pytesseract==0.3.10
pdfkit==1.0.0
streamlit-chat==0.1.0
tqdm==4.66.1
langchain==0.0.308
requests==2.31.0

